# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Tone of the roast
 */
export type RoastStyle = (typeof RoastStyle)[keyof typeof RoastStyle];

export const RoastStyle = {
  friendly: "friendly",
  savage: "savage",
  dark: "dark",
  desi: "desi",
} as const;

export type Job = (typeof Job)[keyof typeof Job];

export const Job = {
  student: "student",
  engineer: "engineer",
  doctor: "doctor",
  designer: "designer",
  unemployed: "unemployed",
  influencer: "influencer",
  other: "other",
} as const;

export type RelationshipStatus =
  (typeof RelationshipStatus)[keyof typeof RelationshipStatus];

export const RelationshipStatus = {
  single: "single",
  in_relationship: "in_relationship",
  married: "married",
  complicated: "complicated",
} as const;

export type Language = (typeof Language)[keyof typeof Language];

export const Language = {
  english: "english",
  hinglish: "hinglish",
  hindi: "hindi",
  spanish: "spanish",
  french: "french",
} as const;

export type ReactionType = (typeof ReactionType)[keyof typeof ReactionType];

export const ReactionType = {
  hilarious: "hilarious",
  savage: "savage",
  dead: "dead",
  too_real: "too_real",
} as const;

export interface ReactionCounts {
  hilarious: number;
  savage: number;
  dead: number;
  too_real: number;
}

export interface Roast {
  id: string;
  text: string;
  style: RoastStyle;
  name: string;
  job: Job;
  city: string;
  weakness?: string | null;
  status: RelationshipStatus;
  language: Language;
  intensity: number;
  reactions: ReactionCounts;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  name: string;
  job: Job;
  text: string;
  burnScore: number;
}

export interface RoastStats {
  totalRoasts: number;
  usersToday: number;
  roastsPerMinute: number;
  worldwideToday: number;
}

export type GenerateRoastBody = {
  /**
   * Person being roasted
   * @minLength 1
   * @maxLength 60
   */
  name: string;
  job: Job;
  /**
   * @minLength 1
   * @maxLength 60
   */
  city: string;
  /** @maxLength 120 */
  weakness?: string | null;
  status: RelationshipStatus;
  language: Language;
  /**
   * Burn intensity from 1 (Baby Roast) to 5 (Nuclear)
   * @minimum 1
   * @maximum 5
   */
  intensity: number;
  style: RoastStyle;
};

export type ReactToRoastBody = {
  type: ReactionType;
};