# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: roasts
    description: AI roast generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /roasts/generate:
    post:
      operationId: generateRoast
      tags: [roasts]
      summary: Generate a personalized AI roast
      description: Send personalized fields and tone preferences; receive a savage AI roast.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 60
                  description: Person being roasted
                job:
                  $ref: "#/components/schemas/Job"
                city:
                  type: string
                  minLength: 1
                  maxLength: 60
                weakness:
                  type: string
                  maxLength: 120
                  nullable: true
                status:
                  $ref: "#/components/schemas/RelationshipStatus"
                language:
                  $ref: "#/components/schemas/Language"
                intensity:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Burn intensity from 1 (Baby Roast) to 5 (Nuclear)
                style:
                  $ref: "#/components/schemas/RoastStyle"
              required:
                - name
                - job
                - city
                - status
                - language
                - intensity
                - style
      responses:
        "200":
          description: Roast generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roast"
  /roasts/trending:
    get:
      operationId: listTrendingRoasts
      tags: [roasts]
      summary: Trending roasts
      description: Returns a rotating list of recent shareable roasts for social proof.
      responses:
        "200":
          description: List of trending roasts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Roast"
  /roasts/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [roasts]
      summary: Hall of Shame leaderboard
      description: Returns the top roasts by burn score for today.
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /roasts/stats:
    get:
      operationId: getRoastStats
      tags: [roasts]
      summary: Roast counters
      description: Returns global counters used as social proof on the landing page.
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoastStats"
  /roasts/{id}:
    get:
      operationId: getRoast
      tags: [roasts]
      summary: Get a roast by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Roast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roast"
        "404":
          description: Not found
  /roasts/{id}/reactions:
    post:
      operationId: reactToRoast
      tags: [roasts]
      summary: React to a roast
      description: Increment a reaction counter on a roast.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  $ref: "#/components/schemas/ReactionType"
              required:
                - type
      responses:
        "200":
          description: Updated reaction counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionCounts"
        "404":
          description: Not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RoastStyle:
      type: string
      enum: [friendly, savage, dark, desi]
      description: Tone of the roast
    Job:
      type: string
      enum: [student, engineer, doctor, designer, unemployed, influencer, other]
    RelationshipStatus:
      type: string
      enum: [single, in_relationship, married, complicated]
    Language:
      type: string
      enum: [english, hinglish, hindi, spanish, french]
    ReactionType:
      type: string
      enum: [hilarious, savage, dead, too_real]
    ReactionCounts:
      type: object
      properties:
        hilarious:
          type: integer
        savage:
          type: integer
        dead:
          type: integer
        too_real:
          type: integer
      required:
        - hilarious
        - savage
        - dead
        - too_real
    Roast:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        style:
          $ref: "#/components/schemas/RoastStyle"
        name:
          type: string
        job:
          $ref: "#/components/schemas/Job"
        city:
          type: string
        weakness:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/RelationshipStatus"
        language:
          $ref: "#/components/schemas/Language"
        intensity:
          type: integer
        reactions:
          $ref: "#/components/schemas/ReactionCounts"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - text
        - style
        - name
        - job
        - city
        - status
        - language
        - intensity
        - reactions
        - createdAt
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        name:
          type: string
        job:
          $ref: "#/components/schemas/Job"
        text:
          type: string
        burnScore:
          type: integer
      required:
        - rank
        - name
        - job
        - text
        - burnScore
    RoastStats:
      type: object
      properties:
        totalRoasts:
          type: integer
        usersToday:
          type: integer
        roastsPerMinute:
          type: integer
        worldwideToday:
          type: integer
      required:
        - totalRoasts
        - usersToday
        - roastsPerMinute
        - worldwideToday