# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Job = (typeof Job)[keyof typeof Job];

export const Job = {
  student: "student",
  engineer: "engineer",
  doctor: "doctor",
  designer: "designer",
  unemployed: "unemployed",
  influencer: "influencer",
  other: "other",
} as const;