# lib/api-zod/src/generated/types/language.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type Language = (typeof Language)[keyof typeof Language];

export const Language = {
  english: "english",
  hinglish: "hinglish",
  hindi: "hindi",
  spanish: "spanish",
  french: "french",
} as const;