# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from "./job";

export interface LeaderboardEntry {
  rank: number;
  name: string;
  job: Job;
  text: string;
  burnScore: number;
}