# lib/api-zod/src/generated/types/reactionType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ReactionType = (typeof ReactionType)[keyof typeof ReactionType];

export const ReactionType = {
  hilarious: "hilarious",
  savage: "savage",
  dead: "dead",
  too_real: "too_real",
} as const;