# lib/api-zod/src/generated/types/relationshipStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type RelationshipStatus =
  (typeof RelationshipStatus)[keyof typeof RelationshipStatus];

export const RelationshipStatus = {
  single: "single",
  in_relationship: "in_relationship",
  married: "married",
  complicated: "complicated",
} as const;