# lib/api-zod/src/generated/types/roast.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from "./job";
import type { Language } from "./language";
import type { ReactionCounts } from "./reactionCounts";
import type { RelationshipStatus } from "./relationshipStatus";
import type { RoastStyle } from "./roastStyle";

export interface Roast {
  id: string;
  text: string;
  style: RoastStyle;
  name: string;
  job: Job;
  city: string;
  weakness?: string | null;
  status: RelationshipStatus;
  language: Language;
  intensity: number;
  reactions: ReactionCounts;
  createdAt: Date;
}